Preserve (current-buffer) while responding to OpenWindow requests.
Any scope protected by with-current-buffer should not call a function
that changes the current buffer if that function isn't _also_
delimited by its _own_ with-current-buffer call.  Here, this principle
was violated by dyalog-ride-open-edit-window; it clobbered the current buffer of
dyalog-ride-receive by invoking pop-to-buffer without protection, which invokes
display-buffer, which modifies the current buffer haphazardly.  This
would lead to fatal out-of-bounds errors where a newly opened source code
window is treated as if it is the process buffer.  By protecting it
like dyalog-ride-open-debugger the window opens successfully.